modify post(/:id) to generate randomName before application sent. modify end route to mark as application
