+++ server/routes/candidateProfile.router.js
@@ -362,97 +362,6 @@ router.delete('/:id', (req, res) => {
 
 
 
-
-
-
-
-router.post('/:id/application', (req, res) => {
-  let randomName ='';
-  let animalList = 'https://www.randomlists.com/data/animals.json';
-  let thingList = 'https://www.randomlists.com/data/things.json'
-
-  const requestAnimal = axios.get(animalList);
-  const requestThing = axios.get(thingList);
-
-
-  axios.all([requestAnimal, requestThing])
-    .then(axios.spread((...response) => {
-
-      const responseAnimal = response[0].data.RandL.items
-      const responseThing = response[1].data.RandL.items
-
-
-      function pickRandom(list) {
-        return list[Math.floor(Math.random() * list.length)];
-      }
-
-      randomName = pickRandom(responseAnimal) + " " + pickRandom(responseThing);
-      console.log(randomName);
-
-
-      const sqlTxt =`INSERT INTO "application" ("candidate_id", "job_post_id", "random_identifier")
-                      VALUES (1, 3, '${randomName}');`;
-
-      
-      // const sqlParams = [
-      //   req.user.user_info.id,
-      //   req.params.id,
-      //   randomName
-      // ];
-
-      pool.query(sqlTxt)
-      .then((response)=>{
-        res.sendStatus(200);
-        console.log('Application successful');
-      })
-
-    })).catch(error => {
-      console.log(error)
-    })
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-})
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
 // router.get('/application/namegenerator', (req, res) => {
 
 //   axios({
